Use `itertools.product` instead of nested loops
Instead of using nested loops and an auxiliary list to generate the
`indexes` tuple, it's probably faster and more concise to just use
`itertools.product` to directly generate the tuple in "one" iteration.
The `if True` blocks have only been inserted to be able to keep the
indentation to keep the diff minimal. That way the actual changes are
not obscured by lots of lines only changing because of being unindented.