Dynamically determine dimension index position
The order in which dimensions are used to index a variable V1 is now
determined dynamically by storing at which position a dimension appears
in the index of a variable in the `d_index` dictionary.
This dictionary is then used by `value_of` to determine the value of a
(different) variable V2 indexed by a subset of the dimensions of V1 at
the index I which should be a full index into V1.
That way we can get the value of 'time_bnds' or 'altitude' matching
specific values of a primary variable (like e.g. 'P') without resorting
to hard coding assumptions about the order in which dimensions appear in
the index of the primary variable.