Drop unnecessary `if 'time' in dims:` check
The time dimension should always be there because we already weeded out
variables which are constant with respect to time earlier. There's a
comment about that above the changed lines. Also, handling dimensions by
computing appropriate values for them like the start/stop timestamps for
time or the grid points for rlat and rlon should be factored out into a
dedicated function anyway. Doing this also makes setting appropriate
defaults for missing dimensions possible.
The commit actually only drops the `if ...` check. It looks bigger
because python forces us to de-dent the whole block but no other lines
were changed, except for joining lines which now fit the 79 character
limit.