Add caching of dimension values
The `DimensionCache` class not only takes care of pre-fetching/creating
dimension values, but also acts as the bridge between the NetCDF format
and how things are stored in the database. That way it can be a central
point for all the messy translation code handling the idiosyncrasies
between the two formats.
The commit is pretty big, but most of it is just cleaned up and
refactored stuff from `import_nc_file`.