Fix a progress assessment and a flush bug
The progress assessment was messed up, as the count starts with zero and
the progress is thus immediately bumped up by one chunk. This is fixed
by starting the count with one.
The second bug was that the last `Value` objects, those which are not
enough to constitute a whole chunk, where not flushed to the database as
the loop just exited. Flushing after the loop takes care of that.