Fix progress bar label alignment
The was erroneous assumption in the original progress bar label
alignment commit, which was that `width` specifies the width of the
padding. It instead specifies the width of the whole interpolated part.
Therefore the variable name does not have to be factored into the
calculation.
Reducing the fixed part by one had to be done because it was
artificially inflated in order to make the case with which the original
commit was tested look right.