Improve import progress bar updates
The amount by which the bar is updated is now calculated instead of
coded as a constant. That way there's only one place which has to be
changed in order to change the chunk.
Obviously updating the bar and emptying the list of mappings had to
swapped as the `len` call would not make sense otherwise.