Prevent `Value`'s foreign keys from being null
Every `Value` instance should have a `timestamp`, a `location` and a
`variable` it belongs to, so the keys by which these are referenced
should never be `NULL`. There can be special `timestamp`s or `location`s
signifying everytime or everywhere, but every `Value` instance should
reference those with appropriate keys.