Enforce uniqueness of `Value`s
There should be exactly one value for each variable at a specific
location at a specific timestamp. This is enforce by putting a compound
uniqueness constraint on the foreign keys referencing these dimensions
in values.
This is also one of the reasons, why not having `NULL`s on these foreign
keys is important.