Allow limiting along dimensions
This commit adds an option that allows limiting dimensions to only
certain indices. This means, that from every variable indexed by the
specified dimensions, only the values which are indexed by the
dimension's values at the specified indexes are imported.
This is actually a special solution, since it doesn't cover limiting
general variables. Hopefully I'll find on for that, but this one should
suffice for now.
Also, the `DimensionCache` is still built for all dimension values, so
limiting doesn't cover this aspect, for now.