Complete the #11 fix
Having the tables already defined in the database leads to some
problems with seemingly conflicting table definitions. Even if they are
actually the same, SQLAlchemy won't notice it. Also, if the database
schema is not reflected, Python won't pick up any already defined
constraints, besides primary keys.
For that reason, the table metadata has to be created with reflection
enabled if we are not dropping annything which is already there. This
also forces us to bind an engine to the metadata object before using it.
Additionally, we have to define how to resolve any perceived conflicts,
which in this case it done by being conservative and setting
`keep_existing` to `True` for all tables, which means already existing
definitions have precedence over new ones.