Import `Float` with module prefix
There will be a need for a `Float` class whose name will clash with the
`Float` datatype from the `sqlalchemy` module. Therefore this patch
removes the direct import of `Float` and adds an import of the whole
`sqlalchemy` module with a shortened name so that the `Float` datatype
can be accessed using the shorthand module prefix.