Try using classical mappings for `Value` JTI
In order to handle flag values, I need to use joined table inheritance
to make the type of the `v` column polymorphic. Unfortunately using the
old scheme of generating the mapped classes for this dynamically doesn't
play well with the `polymorphic_on` mapper argument, so I tried using a
classical mapping.
Now this fails when trying to store a `Value`, as no `id` is generated,
violating a PK constraint. Let's see how we can fix that.