Commit after caching dimensions
If we don't commit after caching the dimensions, caching timestamps
blocks until another process started earlier is done importing it's
variables. This is not what we want, as importing variables doesn't
touch the dimension cache. The reason for this blockage stems from the
fact, that the database doesn't know whether the uniqueness constraint
on timestamps holds, until all transactions touching them have
committed.
Since it's safe to do so, for the reasons outlined above, we better do
it to facilitate importing variables in parallel.