Tick off TODO about variable detection
I don't know if that's all there is to it, but having a 'coordinates'
attribute is at least something that is easily to communicate to data
suppliers and it's more robust than relying on the order of the variable
keys. Which broke down with the new two year dataset anyway... D'Oh.