Stop calling Python script via `subprocess.call`
Importing the script as a module and calling the appropriate functions
directly has the advantage of:

  * being less resource hungry no second Python interpreter needs to be
    started and
  * being a lot more readable.

In order to properly call the function from Python though, the code
actually doing the work had to be moved from the `click.command` to an
unwrapped Python function. It seems that the wrapping `click` decorators
modify the functions in such a way that one can no longer supply the
arguments as written in the function header.
Using an unwrapped Python function has the benefit of being easier to
debug, so it's all good and probably should have been done from the get
go anyway.