Store time series values in chunks
Storing one value per row results in the database being way to big.
To decrease the amount of storage space needed, one row now holds an
array of values instead. This has a pretty large impact, which is why
the commit is so big. I'll try to explain the changes a bit in the
following.

First off, the way timestamps are saved also has to be changed in order
to be in line with timeseries values. That means, that instead of
storing single timestamps, arrays of time segments are now stored. These
segments are represented via two element arrays containing the beginning
and end of each segment respectively. Since the timeseries we are
storing can cover the same time period but in different resolutions,
timespans are no longer uniquely identified by their `start` and `stop`
values but by their resolution to.
Storing arrays of timestamps also means that it doesn't really make
sense to cache the time dimension anymore, but since not doing so
would have been a more invasive change than just changing the caching
code, I ended up still caching the time dimension. The only more complex
change necessary for this to work was the addition of an `exclude`
keyword argument to `DimensionCache.cache`. This argument allows us to
exclude certain parameters of classes mapped to a dimension to be
excluded from being used in the filter that checks whether a dimension
value already exists in the database. This was important since we might
not want to compare all timestamps of a time period in a netCDF file
against values in the database, but instead rely on `start`, `stop` and
`resolution` for performance reasons.
Caching dimensions is something that could use some refactoring though.

Other than that, and besides the obvious needs for changes in the
data model, the only additional thing that had to be changed was how
values are actually inserted into the database. Since we now store
multiple values along the time dimension in one row, we can no longer
use a single index to pull the values out of the netCDF variables, but
we have to use a `slice` instead. Luckily doing so is relatively
straightforward. All we have to do is insert a special case for the time
dimension when generating the tuples used to index the variable and
making sure that we the conversion to float now happens for each value
contained in the sliced. While doing this, I went ahead and rounded the
values to three decimal digits, because according to our project
partners, this is accurate enough for our purposes.
The check for `masked` values doesn't really make sense anymore, too.
And since we now also have a lot less rows to insert, we can drastically
reduce how many chunks to insert at once. It seems like just inserting
one row at a time even seems to be the fastest option, which technically
means we could drop the chunking code, but it was easier to just adjust
the number of chunks for now.

That should be it. Hopefully it explains the changes in this commit in
enough detail for others to follow.