Add `height` to values' unique identifiers
Not doing so lead to false negatives pretty fast, as identical values on
different height levels collided.
The only drawback is that values which are constant wrt. height won't be
checked by the constrained, as these get `Null` value for height.
Might mean I have to move the column to it's own table...