Improve handling the "time" coordinate
The improvement is twofold. First, the name of the `"time"` coordinates
`"bounds"` is no longer hardcoded but extracted from the metadata of the
`"time"` variable. Second, `time` (the Python variable) can now actually
be `None`, a case that didn't really occur with the previous code, at
least not if `"time"` was a variable in the dataset.
This means that variables which are constant wrt. time are now marked
(hopefully correctly). As this addresses a TODO, I removed it, to tick
it off.
Concerning the TODO: unfortunately having a time coordinate of length
one doesn't help to differentiate between a time dependent variable and
one that is constant wrt. time, because after factoring out all other
dimensions, there's usually only one value at the given point in time
and we don't know whether the value is indexed by the time or whether it
is constant. I just hope the other criteria are enough to correctly
detect a variable's time dependence.