Factor out importing a single variable
This code gets a dedicated function now. In order to keep the diff
minimal, I used an additional level on indentation in the new function.
That way one can see that this is really just a change adding two
significant lines.
The cool thing about this though is that importing a single variable
from a NetCDF file is a good unit of work to be done by a single process
and isolating that in a single function makes importing things in
parallel much easier.