Refactor clunky loops
Using comprehensions has the benefit of creating a dedicated scope for
the loop variables so they don't pollute the function's
namespace/scope. This is of special importance when using single letter
variables, as those are prone to collide with other loop variables,
potentially causing hard to track down bugs. I still took the liberty of
using longer variable names to improve readability though.
I also think that comprehensions are more readable than nested loops.

If you ask yourself why the `os.path.isdir(p)` check got removed with
this refactoring: it is because `os.walk` returns an empty list anyway
when called on non-directories, so the check wasn't even necessary in
the first place.