Avoid shadowing in list comprehension
Shadowing the first occurrence/binding of `path` with the second one
works correctly, but reusing a variable name like this might be
confusing and make the code harder to read than necessary. Using a
different name that still correctly conveys the contents of the variable
makes things a lot clearer.