Remove progress bars
These don't make sense doing multiple imports in parallel. That would
just lead to multiple processes writing to the same `stdout` in parallel
which would mess up everything. Since the import will be running at the
scale of months, only having outputs every two ours so is sufficient.

Dropping progress bars means we can remove some variables and arguments
only used for the bars. This includes the `label` when caching
dimensions and the `length` of the progress bar when importing a
variable.

The `if True:` lines are only there to keep the indentation levels the
same as the original code, so that there isn't any noise in the diff for
code blocks that still do the same as before.