Remove `chunk`
The method was only ever used to be able to control the granularity at
which the progress bar was updated and even then it got changed to
chunks of size one a while ago, so the chunking wasn't even used anymore
anyway. The only reason it didn't get removed back then was the fact
that I probably wasn't sure whether I need to use the chunking at a
later point again. I'm pretty sure now, so there's no reason to keep the
code.