Refactor `DimensionCache` to `Dimensions`
Caching is no longer done during `__init__` but in an extra method which
must be called explicitly. This is the first step to dividing this
monolithic class into smaller pieces. Also, with the planned refactoring
to importing in parallel processes, outputting messages from the process
would be problematic from the `__init__` method.