Continuously monitor directories for new files
This means that the logic for submitting jobs to the process pool has to
be moved into the loop checking whether the jobs are finished or not.
While at it, I took the liberty of moving console output out of
`import_nc_file`. I'm submitting the message to the message queue now,
so that most output has the same structure.