Drop fragile detection of constantness wrt. time
The assertions fired in some places where they shouldn't have, so the
heuristics didn't work. So now, we are back detecting things the easy
way via "cell_methods" or specifying it via the command line. Also,
since an absence of a "units" attribute on the "time_bnds" means we have
to copy the one from "time" again, if "time"'s bounds doesn't have one.