Move requirements into `setup.py`
Now that this gets packaged, there's no reason to use `requirements.txt`
anymore. Installing dependencies via `setup.py` is the better way and
gets rid of another file in the repository.
Limiting `oemof.db` to at least version `0.0.7` is done because that's
the version that will fix the outdated pinned `pandas` version and the
issue of `oemof` and `oemof.db` being importable in an arbitrary order.